test(memory): add 42 arena tests and fix telemetry Counter<ulong> to Counter<long> for Metrics compatibility

+++ native/packages/Axrone.Memory/src/Arena/ArenaTelemetryEngine.cs
@@ -5,10 +5,10 @@ namespace Axrone.Memory.Arena;
 internal sealed class ArenaTelemetryEngine : IDisposable
 {
     private readonly Meter _meter;
-    private readonly Counter<ulong> _allocations;
-    private readonly Counter<ulong> _allocatedBytes;
-    private readonly Counter<ulong> _contentionEvents;
-    private readonly Counter<ulong> _commitStalls;
+    private readonly Counter<long> _allocations;
+    private readonly Counter<long> _allocatedBytes;
+    private readonly Counter<long> _contentionEvents;
+    private readonly Counter<long> _commitStalls;
     private readonly Histogram<double> _commitLatencyNs;
     private readonly KeyValuePair<string, object?>[] _tags;
 
@@ -17,10 +17,10 @@ internal sealed class ArenaTelemetryEngine : IDisposable
         _meter = new Meter(meterName, "1.0.0");
         _tags = [new KeyValuePair<string, object?>("arena.instance", instanceName)];
 
-        _allocations = _meter.CreateCounter<ulong>("arena.allocations.count", "{allocations}");
-        _allocatedBytes = _meter.CreateCounter<ulong>("arena.allocated.bytes", "By");
-        _contentionEvents = _meter.CreateCounter<ulong>("arena.contention.cas_cycles", "{cycles}");
-        _commitStalls = _meter.CreateCounter<ulong>("arena.commit.stalls", "{stalls}",
+        _allocations = _meter.CreateCounter<long>("arena.allocations.count", "{allocations}");
+        _allocatedBytes = _meter.CreateCounter<long>("arena.allocated.bytes", "By");
+        _contentionEvents = _meter.CreateCounter<long>("arena.contention.cas_cycles", "{cycles}");
+        _commitStalls = _meter.CreateCounter<long>("arena.commit.stalls", "{stalls}",
             "Commit CAS spin iterations that exhausted the bounded spin budget.");
         _commitLatencyNs = _meter.CreateHistogram<double>("arena.commit.duration", "ns",
             "Wall-clock nanoseconds spent in the ordered CAS commit spin-wait.");
@@ -29,12 +29,12 @@ internal sealed class ArenaTelemetryEngine : IDisposable
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
     public void RecordAllocation(ulong count, nuint bytes)
     {
-        _allocations.Add(count, _tags);
-        _allocatedBytes.Add((ulong)bytes, _tags);
+        _allocations.Add((long)count, _tags);
+        _allocatedBytes.Add((long)bytes, _tags);
     }
 
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    public void RecordContention(ulong cycles) => _contentionEvents.Add(cycles, _tags);
+    public void RecordContention(ulong cycles) => _contentionEvents.Add((long)cycles, _tags);
 
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
     public void RecordCommitStall() => _commitStalls.Add(1, _tags);
